# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AOneLazer Finance API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  totalBalance: number;
  monthlyIncome: number;
  monthlyExpenses: number;
  totalInvestments: number;
  savingsRate: number;
  monthlyChange: number;
  expenseChange: number;
  investmentGain: number;
}

export interface SpendingTrendPoint {
  month: string;
  expenses: number;
  income: number;
  savings: number;
}

export interface CategoryBreakdown {
  category: string;
  amount: number;
  percentage: number;
  color: string;
}

export interface Expense {
  id: number;
  title: string;
  amount: number;
  category: string;
  /** @nullable */
  description?: string | null;
  date: string;
  createdAt: string;
}

export interface CreateExpenseBody {
  title: string;
  amount: number;
  category: string;
  /** @nullable */
  description?: string | null;
  date: string;
}

export interface UpdateExpenseBody {
  title?: string;
  amount?: number;
  category?: string;
  /** @nullable */
  description?: string | null;
  date?: string;
}

export interface Investment {
  id: number;
  name: string;
  /** @nullable */
  ticker?: string | null;
  type: string;
  currentValue: number;
  purchaseValue: number;
  /** @nullable */
  shares?: number | null;
  gainLoss: number;
  gainLossPercent: number;
  createdAt: string;
}

export interface CreateInvestmentBody {
  name: string;
  /** @nullable */
  ticker?: string | null;
  type: string;
  currentValue: number;
  purchaseValue: number;
  /** @nullable */
  shares?: number | null;
}

export interface UpdateInvestmentBody {
  name?: string;
  /** @nullable */
  ticker?: string | null;
  type?: string;
  currentValue?: number;
  purchaseValue?: number;
  /** @nullable */
  shares?: number | null;
}

export interface AllocationItem {
  type: string;
  value: number;
  percentage: number;
  color: string;
}

export interface PortfolioSummary {
  totalValue: number;
  totalGainLoss: number;
  totalGainLossPercent: number;
  allocation: AllocationItem[];
}

export interface Transaction {
  id: number;
  title: string;
  amount: number;
  type: string;
  category: string;
  date: string;
  /** @nullable */
  icon?: string | null;
}

export interface Budget {
  id: number;
  category: string;
  limit: number;
  spent: number;
  remaining: number;
  percentage: number;
  color: string;
  createdAt: string;
}

export interface CreateBudgetBody {
  category: string;
  limit: number;
  color: string;
}

export interface UpdateBudgetBody {
  category?: string;
  limit?: number;
  color?: string;
}

export type ListExpensesParams = {
  category?: string;
  startDate?: string;
  endDate?: string;
  limit?: number;
};

export type ListTransactionsParams = {
  limit?: number;
  type?: string;
};