# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AOneLazer Finance API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  title: string;
  amount: number;
  category: string;
  /** @nullable */
  description?: string | null;
  date: string;
  createdAt: string;
}