# lib/api-zod/src/generated/types/investment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AOneLazer Finance API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Investment {
  id: number;
  name: string;
  /** @nullable */
  ticker?: string | null;
  type: string;
  currentValue: number;
  purchaseValue: number;
  /** @nullable */
  shares?: number | null;
  gainLoss: number;
  gainLossPercent: number;
  createdAt: string;
}