# lib/api-zod/src/generated/types/spendingTrendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AOneLazer Finance API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SpendingTrendPoint {
  month: string;
  expenses: number;
  income: number;
  savings: number;
}