# lib/api-zod/src/generated/types/updateBudgetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AOneLazer Finance API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateBudgetBody {
  category?: string;
  limit?: number;
  color?: string;
}